+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### [Unreleased]
+* FIXED: Update functionality when objects has collections with nested collections
 
 ### [2.4.2] - 2023-06-25
 * FIXED: Duplicate collection data to JSON on save when configured case was not used with collection name